Add files via upload
Please note that this doc for meeting notes was made in early October but was never committed. I apologize for the disorganization and not keeping track with what I plan on committing and when. This was one of the first weeks of working on projects. Our group spent the week researching API and data to get started. We had to do a lot of research since all of us are beginners.