Add files via upload
Please Note that this document was made during the week of October 8th to the 11th but was not committed. I apologize for any inconvenience this disorganization may cause. 
During this week we researched Node.JS and MongoDB and created accounts as well as installations of these programs and tried to start testing how to use them.